HRPV2-102 Refactor to move the case when clauses from query to service layer

+++ oe-hr-portal-service-common/src/main/java/com/openenglish/hr/common/dto/ActivitiesOverviewDto.java
@@ -14,7 +14,7 @@ public class ActivitiesOverviewDto {
     private long privateClasses;
     private long completedLessons;
     private long completedUnits;
-    private long practiceHours;
+    private double practiceHours;
     private long levelPassed;
     private double totalHoursUsage;
 }
